+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 * Harden the security of our GitHub Actions Workflows
 
+## Fixed
+
+* Fix option mode detection from Symfony command definition
+
 ## 1.5.0 (2026-05-27)
 
 ### Features